chore(dev): connect custom mcps
